Add account creation and add/update relevant Postman API + tests.
Add the AccountsController and add the routes for account creation and deletion (for testing). Create the models for credentials and users. Add new and get repository functions to the CredentialsRepository and UsersRepository. Update the Postman API for the 'New Account' folder and added many requests and tests for edge cases.